[UI] Phase 5.a: Split performance/index into focused siblings (#18660)
components/performance/index.tsx was 1203 significant lines, over the
1000-line hard ceiling tracked by `npm run audit:size` (§7.4 of the
restructure plan). Extract four sibling React components and two helper
modules under the existing `ui/components/performance/` folder. The
entry point keeps the same default export and same external interface;
only internal organization changes.

New siblings (kebab-case for helper modules, PascalCase for components,
matching nearby conventions):
  - performance-helpers.tsx (72 sig lines): the `generatePerformanceProfile`
    helper, plus the load-generator constants and info-tooltip JSX.
  - performance-handlers.ts (127): `createPerformanceFormChangeHandler`,
    the curried form-change dispatcher (was the inline `handleChange`).
  - PerformanceForm.tsx (422): the full ModalBody form fields, the
    Advanced Options accordion, and the load-generator radio group.
  - PerformanceFormActions.tsx (79): the four ModalFooter buttons
    (Clear / Results / Save Profile / Run Test).
  - PerformanceCharts.tsx (78): the three optional telemetry chart
    sections rendered below the form.
  - PerformanceTestResults.tsx (46): the post-run results view
    (was the inline `Results` component).

After the split:
  - components/performance/index.tsx is 588 sig lines (under the 600
    warn threshold).
  - over_hard drops from 4 to 3; over_warn drops from 12 to 11.

eslint.config.js: drop components/performance/index.tsx from
legacyMaxLineOffenders (it is no longer over 1000 lines); allowlist
PerformanceForm.tsx for legacyLiteralColorOffenders (it now holds the
inherited `#929292` hex literals); allowlist the five new component
sibling files for legacyInlineStyleOffenders (they inherit inline
style={{}} usage from the original).

Behavior is preserved — same Redux state, same prop interface, same
rendered DOM. The `generatePerformanceProfile` named export is
re-exported from the new helper module so any legacy callers that
imported it from the entry point keep working.

AUDIT size over_warn=11 over_hard=3

Refs #18660
Refs #18654
